requirements.txt
mcp>=1.2.0
httpx>=0.26.0
python-dotenv>=1.0.0 